requirements.txt
streamlit==1.32.0
numpy==1.24.3
nibabel==5.1.0
Pillow==10.0.0
pydicom==2.4.3
torch==2.1.0
torchvision==0.16.0
scikit-image==0.21.0
matplotlib==3.7.1 